src/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
black==23.11.0
certifi==2023.7.22
click==8.1.7
coverage==7.3.2
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.1
flake8==6.1.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.1
httptools==0.6.1
httpx==0.25.1
idna==3.4
influxdb-client==1.38.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
orjson==3.9.10
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
pycodestyle==2.11.1
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyflakes==3.1.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
reactivex==4.0.4
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
